Convert CLI commands to Typescript
All CLI commands now run as typescript files.
