+++ package.json
@@ -26,9 +26,9 @@
   },
   "dependencies": {
     "commander": "^8.3.0",
-    "glob": "^7.2.0",
     "partition-all": "^1.0.1",
     "promise-queue": "^2.2.5",
+    "tiny-glob": "^0.2.9",
     "toml": "^3.0.0",
     "typescript": "^4.4.4",
     "webdriverio": "^7.16.13"
