+++ remote_metadata_summary.sh
@@ -16,7 +16,7 @@ remote_screenshot_dir="${2/%\//}"
 # If copying all the data takes too long at some point,
 # consider a non-temporary directory
 rsync --recursive "$1:$remote_screenshot_dir/" $local_metadata_dir
-./metadata_summary.js --screenshotPath $local_metadata_dir
+npx ts-node metadata_summary.js --screenshotPath $local_metadata_dir
 scp -Cq "$local_metadata_dir/metadata_summary.json" "$1:$2"
 metadata_file_count=$(find $local_metadata_dir -name metadata.json | wc -l)
 echo "Generated summary for $metadata_file_count files"
