+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "outDir": "./built",
     "allowJs": true,
-    "target": "es5"
+    "target": "es2020",
+	"esModuleInterop": true
   },
   "include": [
 	  "./src/**/*",
